# Conflicts:
#	app/src/main/java/org/codethechange/culturemesh/PostsFrag.java
#	app/src/main/java/org/codethechange/culturemesh/RVAdapter.java
#	app/src/main/java/org/codethechange/culturemesh/models/FeedItem.java

+++ app/src/main/java/org/codethechange/culturemesh/models/Post.java
@@ -13,10 +13,10 @@ import java.math.BigInteger;
 @Entity
 public class Post extends FeedItem implements Serializable{
     @PrimaryKey
-    public int id;
+    public long id;
     //When saved in database, we use these.
-    public int userId;
-    public int networkId;
+    public long userId;
+    public long networkId;
 
     public String content;
     public String imgLink;
